move new entity id dto from separate namespace (bc break)
