+++ src/CommandBus/ICommandBus.php
@@ -3,10 +3,6 @@ declare(strict_types = 1);
 
 namespace Damejidlo\CommandBus;
 
-use Damejidlo\CommandBus\Implementation\NewEntityId;
-
-
-
 interface ICommandBus
 {
