+++ src/CommandBus/NewEntityId.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types = 1);
 
-namespace Damejidlo\CommandBus\Implementation;
+namespace Damejidlo\CommandBus;
 
 class NewEntityId
 {
